+++ ntex-net/CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## [3.9.1] - 2026-04-07
+
+* Update tokio compat impl
+
+## [3.9.0] - 2026-04-02
+
+* Update to ntex-error 2.0
+
 ## [3.8.0] - 2026-03-08
 
 * Use ntex_error::Error for connect service